se depuro el codigo php de todo el modulo tareas (edit,eliminar,estado,list,update) //validacione de bases de datos/validacion de request/ mensajes de error

+++ php/tarea/tareas_eliminar.php
@@ -1,22 +1,48 @@
 <?php
 
 require ("../../config/conexion.php");
-$codigo=$_POST["codigo"];
 
 $cnn=conectarDataBase();
 
-$sql="DELETE FROM tarea where id = ?";
+if ($_SERVER["REQUEST_METHOD"]!=="POST") {
+  mostrarJson("error","Metodo no encontrado");
+}
 
-$smt = $cnn->prepare($sql);
+if (!empty($_POST["codigo"])) {
 
-$smt->bind_param("i",$codigo);
+    $codigo=intval($_POST["codigo"]);
 
-$smt->execute();
+    $sql="DELETE FROM tarea where id = ?";
 
-if($smt->affected_rows>0){
+    $smt = $cnn->prepare($sql);
 
-echo "se elimino correctamente";
+    if (!$smt) {
+        mostrarJson("error","No se a preparado la consulta");
+    }
+
+    $smt->bind_param("i",$codigo);
+
+    if ($smt->execute()) {
+        if($smt->affected_rows>0){
+
+        mostrarJson("success","Se a eliminado correctamente");
+    
+        }else{
+            mostrarJson("warning","No se a realizado ningun cambios");
+        }
+    }else{
+        mostrarJson("error","No se a Ejecutado la consulta");
+    }
 
 }else{
-echo "Problemas con el registro";
+    mostrarJson("error","Faltan completar Campos");
+}
+
+function mostrarJson ($status,$message,$extra=[]){
+    header('Content-Type: application/json');
+    echo  json_encode(array_merge([
+         "status" => $status,
+         "message" => $message
+
+    ],$extra));
 }
